pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "parallel-flow-cells"
authors = [
    {name = "Flynn Marshall", email = "flyma@dtu.dk"}
]

description = "Python package to run parallel flow cell experiments by serial communcation with custom PCBs and MFCs."
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests > 2.26.0",
    "pandas",
    "numpy",
    "pyserial",
    "ipython",
    "datetime",
    "matplotlib",
    "pyyaml",
]

[project.optional-dependencies]
lint = [
    "mypy",
]

[project.urls]
repository = "https://github.com/flynn-98/parallel-flow-cells.git"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
disallow_untyped_defs = true